feat(fontend): fixed the bug where players who have left are not removed from state
The client side state was keeping all players in the session. Forgot to implement removal of elemets from the this.players map when they are no longer in the list of
players sent by the backed.

The implementation is a bit tedious because of the aim to reduce the amounts of DOM Tree modifications Angular makes. If the array were to be simply set to the value sent by the backnd it would trigger re-renders every for every update even if no changes were present.